task:4748 - Refactor multiple REST Controller interfaces & impls to a single By-Resource REST Controller
