--- java/src/main/java/com/amido/stacks/workloads/menu/api/v1/QueryMenuController.java
@@ -1,67 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1;
-
-import com.amido.stacks.core.api.annotations.ReadAPIResponses;
-import com.amido.stacks.core.api.annotations.SearchAPIResponses;
-import com.amido.stacks.workloads.menu.api.v1.dto.response.MenuDTO;
-import com.amido.stacks.workloads.menu.api.v1.dto.response.SearchMenuResult;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.Schema;
-import io.swagger.v3.oas.annotations.responses.ApiResponse;
-import java.io.IOException;
-import java.util.UUID;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestAttribute;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RequestParam;
-
-@RequestMapping(
-    path = "/v1/menu",
-    produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8",
-    method = RequestMethod.GET)
-public interface QueryMenuController {
-
-  @GetMapping
-  @Operation(
-      tags = "Menu",
-      summary = "Get or search a list of menus",
-      description =
-          "By passing in the appropriate options, you can search for available menus in the system")
-  @ApiResponse(
-      responseCode = "200",
-      description = "Success",
-      content =
-          @Content(
-              mediaType = "application/json",
-              schema = @Schema(implementation = SearchMenuResult.class)))
-  @SearchAPIResponses
-  ResponseEntity<SearchMenuResult> searchMenu(
-      @RequestParam(value = "searchTerm", required = false) String searchTerm,
-      @RequestParam(value = "restaurantId", required = false) UUID restaurantId,
-      @RequestParam(value = "pageSize", required = false, defaultValue = "20") Integer pageSize,
-      @RequestParam(value = "pageNumber", required = false, defaultValue = "1") Integer pageNumber)
-      throws IOException;
-
-  @GetMapping(value = "/{id}")
-  @Operation(
-      tags = "Menu",
-      summary = "Get a menu",
-      description =
-          "By passing the menu id, you can get access to available categories and items in the menu")
-  @ApiResponse(
-      responseCode = "200",
-      description = "Menu",
-      content =
-          @Content(
-              mediaType = MediaType.APPLICATION_JSON_VALUE,
-              schema = @Schema(implementation = MenuDTO.class)))
-  @ReadAPIResponses
-  ResponseEntity<MenuDTO> getMenu(
-      @PathVariable(name = "id") UUID id,
-      @Parameter(hidden = true) @RequestAttribute("CorrelationId") String correlationId);
-}
