--- java/src/main/java/com/amido/stacks/workloads/menu/api/v1/UpdateMenuController.java
@@ -1,36 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1;
-
-import com.amido.stacks.core.api.annotations.UpdateAPIResponses;
-import com.amido.stacks.core.api.dto.response.ResourceUpdatedResponse;
-import com.amido.stacks.workloads.menu.api.v1.dto.request.UpdateMenuRequest;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.Parameter;
-import java.util.UUID;
-import javax.validation.Valid;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestAttribute;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-
-@RequestMapping(
-    path = "/v1/menu/{id}",
-    consumes = MediaType.APPLICATION_JSON_VALUE,
-    produces = MediaType.APPLICATION_JSON_VALUE + "; charset=utf-8",
-    method = RequestMethod.PUT)
-public interface UpdateMenuController {
-
-  @PutMapping(consumes = "application/json", produces = "application/json; charset=utf-8")
-  @Operation(
-      tags = "Menu",
-      summary = "Update a menu",
-      description = "Update a menu with new information")
-  @UpdateAPIResponses
-  ResponseEntity<ResourceUpdatedResponse> updateMenu(
-      @Parameter(description = "Menu id", required = true) @PathVariable("id") UUID menuId,
-      @Valid @RequestBody UpdateMenuRequest body,
-      @Parameter(hidden = true) @RequestAttribute("CorrelationId") String correlationId);
-}
