+++ java/src/test/java/com/amido/stacks/workloads/menu/api/v1/CategoryControllerTest.java
@@ -1,21 +1,27 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
+package com.amido.stacks.workloads.menu.api.v1;
 
 import static com.amido.stacks.workloads.menu.domain.CategoryHelper.createCategories;
 import static com.amido.stacks.workloads.menu.domain.CategoryHelper.createCategory;
+import static com.amido.stacks.workloads.menu.domain.ItemHelper.createItem;
+import static com.amido.stacks.workloads.menu.domain.MenuHelper.createMenu;
 import static com.azure.cosmos.implementation.Utils.randomUUID;
 import static java.util.UUID.fromString;
 import static org.assertj.core.api.BDDAssertions.then;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.CONFLICT;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.OK;
 
 import com.amido.stacks.core.api.dto.ErrorResponse;
+import com.amido.stacks.core.api.dto.response.ResourceCreatedResponse;
 import com.amido.stacks.core.api.dto.response.ResourceUpdatedResponse;
 import com.amido.stacks.workloads.Application;
+import com.amido.stacks.workloads.menu.api.v1.dto.request.CreateCategoryRequest;
 import com.amido.stacks.workloads.menu.api.v1.dto.request.UpdateCategoryRequest;
 import com.amido.stacks.workloads.menu.domain.Category;
 import com.amido.stacks.workloads.menu.domain.Menu;
@@ -25,6 +31,7 @@ import com.amido.stacks.workloads.util.TestHelper;
 import com.azure.spring.autoconfigure.cosmos.CosmosAutoConfiguration;
 import com.azure.spring.autoconfigure.cosmos.CosmosHealthConfiguration;
 import com.azure.spring.autoconfigure.cosmos.CosmosRepositoriesAutoConfiguration;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
@@ -53,15 +60,154 @@ import org.springframework.test.context.ActiveProfiles;
     })
 @Tag("Integration")
 @ActiveProfiles("test")
-class UpdateCategoryControllerImplTest {
+public class CategoryControllerTest {
+  public static final String CREATE_CATEGORY = "%s/v1/menu/%s/category";
 
   public static final String UPDATE_CATEGORY = "%s/v1/menu/%s/category/%s";
+
+  public static final String DELETE_CATEGORY = "%s/v1/menu/%s/category/%s";
 
   @LocalServerPort private int port;
 
   @Autowired private TestRestTemplate testRestTemplate;
 
   @MockBean private MenuRepository menuRepository;
+
+  @Test
+  void testCanNotAddCategoryIfMenuNotPresent() {
+    // Given
+    UUID menuId = randomUUID();
+    when(menuRepository.findById(eq(menuId.toString()))).thenReturn(Optional.empty());
+
+    CreateCategoryRequest request =
+        new CreateCategoryRequest("test Category Name", "test Category Description");
+
+    // When
+    var response =
+        this.testRestTemplate.postForEntity(
+            String.format(CREATE_CATEGORY, TestHelper.getBaseURL(port), menuId),
+            request,
+            ErrorResponse.class);
+
+    // Then
+    then(response).isNotNull();
+    then(response.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
+  }
+
+  @Test
+  void testNoDescriptionGivenReturnsBadRequest() {
+    // Given
+    CreateCategoryRequest request = new CreateCategoryRequest("test Category Name", "");
+
+    // When
+    var response =
+        this.testRestTemplate.postForEntity(
+            String.format(CREATE_CATEGORY, TestHelper.getBaseURL(port), randomUUID()),
+            request,
+            ErrorResponse.class);
+
+    // Then
+    then(response).isNotNull();
+    then(response.getStatusCode()).isEqualTo(BAD_REQUEST);
+    then(response.getBody().getDescription())
+        .isEqualTo("Invalid Request: {description=must not be blank}");
+  }
+
+  @Test
+  void testInvalidMenuIdWilThrowBadRequest() {
+    // Given
+    CreateCategoryRequest request = new CreateCategoryRequest("test Category Name", "");
+
+    // When
+    var response =
+        this.testRestTemplate.postForEntity(
+            String.format(CREATE_CATEGORY, TestHelper.getBaseURL(port), "XXXXXX"),
+            request,
+            ErrorResponse.class);
+
+    // Then
+    then(response).isNotNull();
+    then(response.getStatusCode()).isEqualTo(BAD_REQUEST);
+  }
+
+  @Test
+  void testAddCategory() {
+    // Given
+    Menu menu = MenuHelper.createMenu(1);
+    when(menuRepository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
+    when(menuRepository.save(any(Menu.class))).thenReturn(menu);
+
+    CreateCategoryRequest request =
+        new CreateCategoryRequest("test Category Name", "test Category Description");
+
+    // When
+    var response =
+        this.testRestTemplate.postForEntity(
+            String.format(CREATE_CATEGORY, TestHelper.getBaseURL(port), menu.getId()),
+            request,
+            ResourceCreatedResponse.class);
+
+    // Then
+    ArgumentCaptor<Menu> captor = ArgumentCaptor.forClass(Menu.class);
+    verify(menuRepository, times(1)).save(captor.capture());
+    Menu created = captor.getValue();
+
+    then(created.getName()).isEqualTo(menu.getName());
+    then(created.getDescription()).isEqualTo(menu.getDescription());
+    then(created.getCategories().size()).isEqualTo(1);
+    then(created.getCategories().get(0).getName()).isEqualTo(request.getName());
+    then(created.getCategories().get(0).getDescription()).isEqualTo(request.getDescription());
+
+    then(response).isNotNull();
+    then(response.getBody()).isNotNull();
+    then(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
+    then(created.getCategories().get(0).getId()).isEqualTo(response.getBody().getId().toString());
+  }
+
+  @Test
+  void testCannotAddCategoryWhichAlreadyExists() {
+    // Given
+    Menu menu = MenuHelper.createMenu(1);
+    Category category =
+        new Category(
+            UUID.randomUUID().toString(), "cat name", "cat description", new ArrayList<>());
+    menu.addOrUpdateCategory(category);
+
+    when(menuRepository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
+
+    CreateCategoryRequest request =
+        new CreateCategoryRequest(category.getName(), "test Category Description");
+
+    // When
+    var response =
+        this.testRestTemplate.postForEntity(
+            String.format(CREATE_CATEGORY, TestHelper.getBaseURL(port), menu.getId()),
+            request,
+            ErrorResponse.class);
+
+    // Then
+    then(response).isNotNull();
+    then(response.getStatusCode()).isEqualTo(HttpStatus.CONFLICT);
+  }
+
+  @Test
+  void testNoNameGivenReturnsBadRequest() {
+    // Given
+    CreateCategoryRequest request = new CreateCategoryRequest("", "Some description");
+
+    // When
+    var response =
+        this.testRestTemplate.postForEntity(
+            String.format(CREATE_CATEGORY, TestHelper.getBaseURL(port), randomUUID()),
+            request,
+            ErrorResponse.class);
+
+    // Then
+    then(response).isNotNull();
+    then(response.getStatusCode()).isEqualTo(BAD_REQUEST);
+    then(response.getBody().getDescription())
+        .isEqualTo("Invalid Request: {name=must not be blank}");
+  }
 
   @Test
   void testUpdateCategorySuccess() {
@@ -317,5 +463,109 @@ class UpdateCategoryControllerImplTest {
     then(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
     then(response.getBody().getDescription())
         .isEqualTo("Invalid Request: {description=must not be blank}");
+  }
+
+  @Test
+  void testDeleteCategorySuccess() {
+    // Given
+    Menu menu = createMenu(1);
+    Category category = createCategory(0);
+    menu.setCategories(List.of(category));
+    when(menuRepository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
+
+    // When
+    String requestUrl =
+        String.format(DELETE_CATEGORY, TestHelper.getBaseURL(port), menu.getId(), category.getId());
+    var response =
+        this.testRestTemplate.exchange(
+            requestUrl,
+            HttpMethod.DELETE,
+            new HttpEntity<>(TestHelper.getRequestHttpEntity()),
+            ErrorResponse.class);
+
+    // Then
+    verify(menuRepository, times(1)).save(menu);
+    then(response.getStatusCode()).isEqualTo(OK);
+    Optional<Menu> optMenu = menuRepository.findById(menu.getId());
+    Menu updated = optMenu.get();
+    then(updated.getCategories()).isNotNull();
+  }
+
+  @Test
+  void testDeleteCategoryWithAnItem() {
+    // Given
+    Menu menu = createMenu(1);
+    Category category = createCategory(0);
+    category.addOrUpdateItem(createItem(0));
+    menu.addOrUpdateCategory(category);
+
+    when(menuRepository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
+
+    // When
+    String requestUrl =
+        String.format(DELETE_CATEGORY, TestHelper.getBaseURL(port), menu.getId(), category.getId());
+    var response =
+        this.testRestTemplate.exchange(
+            requestUrl,
+            HttpMethod.DELETE,
+            new HttpEntity<>(TestHelper.getRequestHttpEntity()),
+            ErrorResponse.class);
+
+    // Then
+    then(response.getStatusCode()).isEqualTo(OK);
+    ArgumentCaptor<Menu> captor = ArgumentCaptor.forClass(Menu.class);
+    verify(menuRepository, times(1)).save(captor.capture());
+    Menu updatedMenu = captor.getValue();
+    then(updatedMenu.getCategories()).isNotNull();
+  }
+
+  @Test
+  void testDeleteCategoryWithInvalidCategoryId() {
+    // Given
+    Menu menu = createMenu(1);
+    Category category = createCategory(0);
+    menu.setCategories(List.of(category));
+    when(menuRepository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
+
+    // When
+    String requestUrl =
+        String.format(DELETE_CATEGORY, TestHelper.getBaseURL(port), menu.getId(), menu.getId());
+    var response =
+        this.testRestTemplate.exchange(
+            requestUrl,
+            HttpMethod.DELETE,
+            new HttpEntity<>(TestHelper.getRequestHttpEntity()),
+            ErrorResponse.class);
+
+    // Then
+    verify(menuRepository, times(0)).save(menu);
+    then(response.getStatusCode()).isEqualTo(NOT_FOUND);
+  }
+
+  @Test
+  void testDeleteACategoryFromList() {
+    // Given
+    Menu menu = createMenu(1);
+    List<Category> categories = createCategories(2);
+    menu.setCategories(categories);
+    when(menuRepository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
+
+    // When
+    String requestUrl =
+        String.format(
+            DELETE_CATEGORY, TestHelper.getBaseURL(port), menu.getId(), categories.get(0).getId());
+    var response =
+        this.testRestTemplate.exchange(
+            requestUrl,
+            HttpMethod.DELETE,
+            new HttpEntity<>(TestHelper.getRequestHttpEntity()),
+            ErrorResponse.class);
+
+    // Then
+    verify(menuRepository, times(1)).save(menu);
+    then(response.getStatusCode()).isEqualTo(OK);
+    Optional<Menu> byId = menuRepository.findById(menu.getId());
+    Menu updatedMenu = byId.get();
+    then(updatedMenu.getCategories()).hasSize(1);
   }
 }
