--- java/src/test/java/com/amido/stacks/workloads/menu/api/v1/impl/CreateMenuControllerImplTest.java
@@ -1,138 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
-
-import static com.amido.stacks.workloads.menu.domain.MenuHelper.createMenu;
-import static com.amido.stacks.workloads.util.TestHelper.getBaseURL;
-import static java.util.UUID.fromString;
-import static org.assertj.core.api.BDDAssertions.then;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.when;
-
-import com.amido.stacks.core.api.dto.ErrorResponse;
-import com.amido.stacks.core.api.dto.response.ResourceCreatedResponse;
-import com.amido.stacks.workloads.Application;
-import com.amido.stacks.workloads.menu.api.v1.dto.request.CreateMenuRequest;
-import com.amido.stacks.workloads.menu.domain.Menu;
-import com.amido.stacks.workloads.menu.repository.MenuRepository;
-import com.azure.spring.autoconfigure.cosmos.CosmosAutoConfiguration;
-import com.azure.spring.autoconfigure.cosmos.CosmosHealthConfiguration;
-import com.azure.spring.autoconfigure.cosmos.CosmosRepositoriesAutoConfiguration;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.UUID;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpStatus;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest(
-    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
-    classes = Application.class)
-@EnableAutoConfiguration(
-    exclude = {
-      CosmosRepositoriesAutoConfiguration.class,
-      CosmosAutoConfiguration.class,
-      CosmosHealthConfiguration.class
-    })
-@Tag("Integration")
-@ActiveProfiles("test")
-class CreateMenuControllerImplTest {
-
-  public static final String CREATE_MENU = "/v1/menu";
-
-  @LocalServerPort private int port;
-
-  @Autowired private TestRestTemplate testRestTemplate;
-
-  @MockBean private MenuRepository menuRepository;
-
-  @Test
-  void testCreateNewMenu() {
-    // Given
-    Menu m = createMenu(1);
-    CreateMenuRequest request =
-        new CreateMenuRequest(
-            m.getName(), m.getDescription(), UUID.fromString(m.getRestaurantId()), m.getEnabled());
-
-    when(menuRepository.findAllByRestaurantIdAndName(
-            eq(m.getRestaurantId()), eq(m.getName()), any(Pageable.class)))
-        .thenReturn(new PageImpl<>(Collections.emptyList()));
-    when(menuRepository.save(any(Menu.class))).thenReturn(m);
-
-    // When
-    var response =
-        this.testRestTemplate.postForEntity(
-            getBaseURL(port) + CREATE_MENU, request, ResourceCreatedResponse.class);
-
-    // Then
-    then(response.getStatusCode()).isEqualTo(HttpStatus.CREATED);
-  }
-
-  @Test
-  void testThrowsErrorOnExists() {
-    // Given
-    Menu m = createMenu(1);
-    CreateMenuRequest request =
-        new CreateMenuRequest(
-            m.getName(), m.getDescription(), UUID.fromString(m.getRestaurantId()), m.getEnabled());
-
-    when(menuRepository.findAllByRestaurantIdAndName(
-            eq(m.getRestaurantId()), eq(m.getName()), any(Pageable.class)))
-        .thenReturn(new PageImpl<>(Arrays.asList(m)));
-
-    // When
-    var response =
-        this.testRestTemplate.postForEntity(
-            getBaseURL(port) + CREATE_MENU, request, ErrorResponse.class);
-
-    // Then
-    then(response.getStatusCode()).isEqualTo(HttpStatus.CONFLICT);
-  }
-
-  @Test
-  void testWhenDescriptionNotGivenReturnsBadRequest() {
-    // Given
-    Menu m = createMenu(1);
-    CreateMenuRequest request =
-        new CreateMenuRequest(m.getName(), "", fromString(m.getRestaurantId()), m.getEnabled());
-
-    when(menuRepository.save(any(Menu.class))).thenReturn(m);
-    // When
-    var response =
-        this.testRestTemplate.postForEntity(
-            getBaseURL(port) + CREATE_MENU, request, ErrorResponse.class);
-
-    // Then
-    then(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
-    then(response.getBody().getDescription())
-        .isEqualTo("Invalid Request: {description=must not be blank}");
-  }
-
-  @Test
-  void testWhenNoNameReturnsBadRequest() {
-    // Given
-    Menu m = createMenu(1);
-    CreateMenuRequest request =
-        new CreateMenuRequest(
-            "", m.getDescription(), fromString(m.getRestaurantId()), m.getEnabled());
-
-    when(menuRepository.save(any(Menu.class))).thenReturn(m);
-    // When
-    var response =
-        this.testRestTemplate.postForEntity(
-            getBaseURL(port) + CREATE_MENU, request, ErrorResponse.class);
-
-    // Then
-    then(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
-    then(response.getBody().getDescription())
-        .isEqualTo("Invalid Request: {name=must not be blank}");
-  }
-}
