--- java/src/test/java/com/amido/stacks/workloads/menu/api/v1/impl/DeleteMenuControllerImplTest.java
@@ -1,89 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
-
-import static org.assertj.core.api.BDDAssertions.then;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-import static org.springframework.http.HttpStatus.NOT_FOUND;
-import static org.springframework.http.HttpStatus.OK;
-
-import com.amido.stacks.core.api.dto.ErrorResponse;
-import com.amido.stacks.workloads.Application;
-import com.amido.stacks.workloads.menu.domain.Menu;
-import com.amido.stacks.workloads.menu.domain.MenuHelper;
-import com.amido.stacks.workloads.menu.repository.MenuRepository;
-import com.amido.stacks.workloads.util.TestHelper;
-import com.azure.spring.autoconfigure.cosmos.CosmosAutoConfiguration;
-import com.azure.spring.autoconfigure.cosmos.CosmosHealthConfiguration;
-import com.azure.spring.autoconfigure.cosmos.CosmosRepositoriesAutoConfiguration;
-import java.util.Optional;
-import java.util.UUID;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest(
-    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
-    classes = Application.class)
-@EnableAutoConfiguration(
-    exclude = {
-      CosmosRepositoriesAutoConfiguration.class,
-      CosmosAutoConfiguration.class,
-      CosmosHealthConfiguration.class
-    })
-@Tag("Integration")
-@ActiveProfiles("test")
-class DeleteMenuControllerImplTest {
-
-  public static final String DELETE_MENU = "%s/v1/menu/%s";
-
-  @LocalServerPort private int port;
-
-  @Autowired private TestRestTemplate testRestTemplate;
-
-  @MockBean private MenuRepository repository;
-
-  @Test
-  void testDeleteMenuSuccess() {
-    // Given
-    Menu menu = MenuHelper.createMenu(1);
-    when(repository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
-
-    var response =
-        this.testRestTemplate.exchange(
-            String.format(DELETE_MENU, TestHelper.getBaseURL(port), menu.getId()),
-            HttpMethod.DELETE,
-            new HttpEntity<>(TestHelper.getRequestHttpEntity()),
-            ResponseEntity.class);
-    // Then
-    verify(repository, times(1)).delete(menu);
-    then(response.getStatusCode()).isEqualTo(OK);
-  }
-
-  @Test
-  void testDeleteMenuWithInvalidId() {
-    // Given
-    Menu menu = MenuHelper.createMenu(1);
-    when(repository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
-
-    var response =
-        this.testRestTemplate.exchange(
-            String.format(DELETE_MENU, TestHelper.getBaseURL(port), UUID.randomUUID().toString()),
-            HttpMethod.DELETE,
-            new HttpEntity<>(TestHelper.getRequestHttpEntity()),
-            ErrorResponse.class);
-    // Then
-    verify(repository, times(0)).delete(menu);
-    then(response.getStatusCode()).isEqualTo(NOT_FOUND);
-  }
-}
