--- java/src/test/java/com/amido/stacks/workloads/menu/api/v1/impl/UpdateMenuControllerImplTest.java
@@ -1,152 +1,0 @@
-package com.amido.stacks.workloads.menu.api.v1.impl;
-
-import static com.azure.cosmos.implementation.Utils.randomUUID;
-import static org.assertj.core.api.BDDAssertions.then;
-import static org.mockito.ArgumentMatchers.eq;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
-import com.amido.stacks.core.api.dto.ErrorResponse;
-import com.amido.stacks.core.api.dto.response.ResourceUpdatedResponse;
-import com.amido.stacks.workloads.Application;
-import com.amido.stacks.workloads.menu.api.v1.dto.request.UpdateMenuRequest;
-import com.amido.stacks.workloads.menu.domain.Menu;
-import com.amido.stacks.workloads.menu.domain.MenuHelper;
-import com.amido.stacks.workloads.menu.repository.MenuRepository;
-import com.amido.stacks.workloads.util.TestHelper;
-import com.azure.spring.autoconfigure.cosmos.CosmosAutoConfiguration;
-import com.azure.spring.autoconfigure.cosmos.CosmosHealthConfiguration;
-import com.azure.spring.autoconfigure.cosmos.CosmosRepositoriesAutoConfiguration;
-import java.util.Optional;
-import java.util.UUID;
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-import org.mockito.ArgumentCaptor;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.boot.web.server.LocalServerPort;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest(
-    webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
-    classes = Application.class)
-@EnableAutoConfiguration(
-    exclude = {
-      CosmosRepositoriesAutoConfiguration.class,
-      CosmosAutoConfiguration.class,
-      CosmosHealthConfiguration.class
-    })
-@Tag("Integration")
-@ActiveProfiles("test")
-class UpdateMenuControllerImplTest {
-
-  public static final String UPDATE_MENU = "%s/v1/menu/%s";
-
-  @LocalServerPort private int port;
-
-  @Autowired private TestRestTemplate testRestTemplate;
-
-  @MockBean private MenuRepository menuRepository;
-
-  @Test
-  void testUpdateSuccess() {
-    // Given
-    Menu menu = MenuHelper.createMenu(0);
-    when(menuRepository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
-
-    UpdateMenuRequest request = new UpdateMenuRequest("new name", "new description", false);
-
-    // When
-    var response =
-        this.testRestTemplate.exchange(
-            String.format(UPDATE_MENU, TestHelper.getBaseURL(port), menu.getId()),
-            HttpMethod.PUT,
-            new HttpEntity<>(request, TestHelper.getRequestHttpEntity()),
-            ResourceUpdatedResponse.class);
-
-    // Then
-    ArgumentCaptor<Menu> captor = ArgumentCaptor.forClass(Menu.class);
-    verify(menuRepository, times(1)).save(captor.capture());
-    Menu updated = captor.getValue();
-
-    then(updated.getName()).isEqualTo(request.getName());
-    then(updated.getDescription()).isEqualTo(request.getDescription());
-    then(updated.getEnabled()).isEqualTo(request.getEnabled());
-    then(updated.getRestaurantId()).isEqualTo(menu.getRestaurantId());
-
-    then(response).isNotNull();
-    then(response.getStatusCode()).isEqualTo(HttpStatus.OK);
-  }
-
-  @Test
-  void testCannotUpdateIfMenuDoesntExist() {
-    // Given
-    UUID menuId = randomUUID();
-    when(menuRepository.findById(eq(menuId.toString()))).thenReturn(Optional.empty());
-
-    UpdateMenuRequest request = new UpdateMenuRequest("name", "description", true);
-
-    // When
-    var response =
-        this.testRestTemplate.exchange(
-            String.format(UPDATE_MENU, TestHelper.getBaseURL(port), menuId),
-            HttpMethod.PUT,
-            new HttpEntity<>(request, TestHelper.getRequestHttpEntity()),
-            ErrorResponse.class);
-
-    // Then
-    then(response).isNotNull();
-    then(response.getStatusCode()).isEqualTo(HttpStatus.NOT_FOUND);
-  }
-
-  @Test
-  void testUpdateMenuWithNoNameReturnsBadRequest() {
-    // Given
-    Menu menu = MenuHelper.createMenu(0);
-    when(menuRepository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
-
-    UpdateMenuRequest request = new UpdateMenuRequest("", "new description", false);
-
-    // When
-    var response =
-        this.testRestTemplate.exchange(
-            String.format(UPDATE_MENU, TestHelper.getBaseURL(port), menu.getId()),
-            HttpMethod.PUT,
-            new HttpEntity<>(request, TestHelper.getRequestHttpEntity()),
-            ErrorResponse.class);
-
-    // Then
-    then(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
-    then(response.getBody().getDescription())
-        .isEqualTo("Invalid Request: {name=must not be blank}");
-  }
-
-  @Test
-  void testUpdateMenuWithNoDescriptionReturnsBadRequest() {
-    // Given
-    Menu menu = MenuHelper.createMenu(0);
-    when(menuRepository.findById(eq(menu.getId()))).thenReturn(Optional.of(menu));
-
-    UpdateMenuRequest request = new UpdateMenuRequest("Updated Name", "", false);
-
-    // When
-    var response =
-        this.testRestTemplate.exchange(
-            String.format(UPDATE_MENU, TestHelper.getBaseURL(port), menu.getId()),
-            HttpMethod.PUT,
-            new HttpEntity<>(request, TestHelper.getRequestHttpEntity()),
-            ErrorResponse.class);
-
-    // Then
-    then(response.getStatusCode()).isEqualTo(HttpStatus.BAD_REQUEST);
-    then(response.getBody().getDescription())
-        .isEqualTo("Invalid Request: {description=must not be blank}");
-  }
-}
